COMMERCE-6528 batch-engine - move to new package
